Refactors DataSet model
Organizes DataSet model in modules:

- DCAT::V2::DataSet for v2 spec
- DCAT::Commons for shared methods between versions
